requirements.txt
pytest~=9.0.2
selenium~=4.41.0
allure-pytest
requests
allure-python-commons~=2.15.3